Moved code to its correct location
